version before insane

+++ Victor-Pidlatyuk/app/Http/Controllers/OAuthController.php
@@ -7,22 +7,22 @@ use Illuminate\Support\Facades\Http;
 
 class OAuthController extends Controller
 {
-    public function redirect(Request $request)
+    public function redirect()
     {
         return redirect(env('OAUTH_SERVER_URI') . '/api/oauth/authorize?redirect_uri='.env('REDIRECT_URI'));
     }
 
     public function callback(Request $request)
     {
+        if($request->get('accept') == 'false') return redirect('/');
+
         $response = Http::post(config('services.oauth_server.uri') . '/oauth/token', [
             'grant_type' => 'authorization_code',
             'client_id' => config('services.oauth_server.client_id'),
             'client_secret' => config('services.oauth_server.client_secret'),
             'redirect_uri' => config('services.oauth_server.redirect'),
             'code' => $request->code
-        ]);
-
-        $response = $response->json();
+        ])->json();
 
         $request->user()->token()->delete();
